[ADD/EDIT] Change Sigmoid, and generate prediction
Changing the sigmoid to rely on the fact that np.exp will operate
elementwise, this actually simplifies the function a great deal.

Creating outputter function to write prediction to a file of choice.
This is done prediction by prediction to deal with the fact that
ypredictions is an array of 1-element arrays.